v1.5.4 - 26.2-fabric: settings reset button, terrain classifier height overflow, malformed-entry resilience, contour interpolation fix, hue bar, render-complete marker, private icon copies, volatile/teardown hardening, analysis button default on

+++ src/main/java/caeruleusTait/world/preview/client/gui/screens/PreviewContainer.java
@@ -2535,7 +2535,7 @@ public void onScreenReentry() {
         // noise sampling for that pixel.
         final PreviewData.BiomeData[] biomeDataById = previewData.biomeId2BiomeData();
         final TerrainCategory[] categoryById = new TerrainCategory[biomeDataById.length];
-        final byte[] estimatedHeightById = new byte[biomeDataById.length];
+        final short[] estimatedHeightById = new short[biomeDataById.length];
         try {
             Registry<Biome> biomeRegistry = workManager.worldgenContext().registryAccess()
                     .compositeAccess().lookupOrThrow(Registries.BIOME);
